Optimize the implementation of _cairo_lzw_compress.
This switches the implementation of _cairo_lzw_compress from using
cairo-hash.c to a custom implementation of the same algorithm that
takes advantage of the specific details of what is needed here.

This final version now has comparable performance to the libtiff code,
but should hopefully be easier to understand.

Add a bunch of comments explaining how the various pieces work.