Support multiple glyph image formats within the same font (fix #4705).
If we find a glyph image of a format that is different from the font's
format, then we use a temporary image and convert to the font's
format.

This provides a simple fix for the crash identified in bug #4705:

	crash at XRenderAddGlyphs
	https://bugs.freedesktop.org/show_bug.cgi?id=4705

A later improvement would guarantee that the "font's format" was
carefully chosen to be able to contain all glyphs in the font without
information loss.