Add skip list implementation (many thanks to Keith Packard)
The files here are copied directly from the standalone skiplist module
available from:

	git clone git://cworth.org/~cworth/skiplist

In particular the files come from the double branch and the following
commit on that branch:

	8b5a439c68e220cf1514d9b3141a1dbdce8af585

Also of interest is the original skiplist module hosted by Keith Packard
that is the original implementation on which these files were based.
Since the cworth/skiplist branched off of keithp's, Keith has also
now implemented a doubly-linked variant which might be interesting for
further simplification of the code. See:

	git clone git://keithp.com/git/skiplist

and the double-link branch there.